+++ console_gpt/menus/command_handler.py
@@ -3,8 +3,8 @@ from typing import Optional
 from console_gpt.custom_stdout import custom_print
 from console_gpt.general_utils import help_message
 from console_gpt.mcp_client import shutdown
-from console_gpt.menus.tools_menu import tools_menu
 from console_gpt.menus.settings_menu import settings_menu
+from console_gpt.menus.tools_menu import tools_menu
 from console_gpt.prompts.file_prompt import file_prompt
 from console_gpt.prompts.image_prompt import upload_image
 from console_gpt.prompts.multiline_prompt import multiline_prompt